gather exp/pi circuits tests
