+++ circuit-benchmarks/src/evm_circuit.rs
@@ -20,7 +20,7 @@ mod evm_circ_benches {
     use rand::SeedableRng;
     use rand_xorshift::XorShiftRng;
     use std::env::var;
-    use zkevm_circuits::evm_circuit::{witness::block_convert, EvmCircuit};
+    use zkevm_circuits::evm_circuit::{witness::block_convert, TestEvmCircuit};
 
     #[cfg_attr(not(feature = "benches"), ignore)]
     #[test]
@@ -52,7 +52,7 @@ mod evm_circ_benches {
 
         let block = block_convert(&builder.block, &builder.code_db).unwrap();
 
-        let circuit = EvmCircuit::<Fr>::new(block);
+        let circuit = TestEvmCircuit::<Fr>::new(block);
         let mut rng = XorShiftRng::from_seed([
             0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06,
             0xbc, 0xe5,
@@ -83,7 +83,7 @@ mod evm_circ_benches {
             Challenge255<G1Affine>,
             XorShiftRng,
             Blake2bWrite<Vec<u8>, G1Affine, Challenge255<G1Affine>>,
-            EvmCircuit<Fr>,
+            TestEvmCircuit<Fr>,
         >(
             &general_params,
             &pk,
