+++ circuit-benchmarks/src/packed_multi_keccak.rs
@@ -17,7 +17,7 @@ mod tests {
     use rand::SeedableRng;
     use rand_xorshift::XorShiftRng;
     use std::env::var;
-    use zkevm_circuits::keccak_circuit::KeccakCircuit;
+    use zkevm_circuits::keccak_circuit::TestKeccakCircuit;
 
     #[cfg_attr(not(feature = "benches"), ignore)]
     #[test]
@@ -37,7 +37,7 @@ mod tests {
         let inputs = vec![(0u8..135).collect::<Vec<_>>(); 3];
 
         // Create the circuit. Leave last dozens of rows for blinding.
-        let circuit = KeccakCircuit::new(2usize.pow(degree) - 64, inputs);
+        let circuit = TestKeccakCircuit::new(2usize.pow(degree) - 64, inputs);
 
         // Initialize the polynomial commitment parameters
         let mut rng = XorShiftRng::from_seed([
@@ -70,7 +70,7 @@ mod tests {
             Challenge255<G1Affine>,
             XorShiftRng,
             Blake2bWrite<Vec<u8>, G1Affine, Challenge255<G1Affine>>,
-            KeccakCircuit<Fr>,
+            TestKeccakCircuit<Fr>,
         >(
             &general_params,
             &pk,
