+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -8,3 +8,4 @@ pub(crate) mod param;
 /// Bytecode circuit tester
 #[cfg(any(feature = "test", test, feature = "test-circuits"))]
 pub mod test;
+pub use test::BytecodeCircuit as TestBytecodeCircuit;
